reorganize shared example specs
