+++ spec/support/actor_example_class.rb
